requirements.txt
Flask
waitress
openai-whisper
torch
numpy